+++ common/utilities/src/main/java/is/codion/common/utilities/logging/LoggerProxy.java
@@ -88,7 +88,7 @@ public interface LoggerProxy {
 			return NONE;
 		}
 		catch (ServiceConfigurationError e) {
-			throw Exceptions.runtime(e.getCause());
+			throw Exceptions.runtime(e, ServiceConfigurationError.class);
 		}
 	}
 }