+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/scaler/Scaler.java
@@ -78,7 +78,7 @@ public interface Scaler {
 							.findFirst();
 		}
 		catch (ServiceConfigurationError e) {
-			throw Exceptions.runtime(e.getCause());
+			throw Exceptions.runtime(e, ServiceConfigurationError.class);
 		}
 	}
 }